Refactor image metadata csv creation.
Refactor logic to create a csv file from .eif and sbet data that can be
imported to Agisoft as references for the images.

CSV Header order now matches the default for Agisoft. The script accepts two
types of .eif files. One only having the timestamp for the images and getting
IMU data from SBET and the other where all information is provided by the .eif
and only needs to be consolidated and brought to the Agisfot column order.

Script itself also uses input parameters now for base path to the needed files.