Images time table - Fix precision for IMU values.
Parsing the SBET csv used float32 for each column, which resulted in loss of
precision. Additionally using a float parser introduced errors in precision as
well.
Changing to do the radians conversion already when transforming the binary SBET
file to a CSV. Also only parse the GPS time as float and use Decimal type for
all other columns when reading the CSV file into pandas to query it.